requirements.txt
streamlit
faiss-cpu
sentence-transformers
google-genai
numpy
PyMuPDF
nltk
tqdm
